Update `physical_plan` tests to not use SessionContext
