+++ datafusion/core/src/physical_plan/coalesce_partitions.rs
@@ -176,7 +176,6 @@ mod tests {
 
     use super::*;
     use crate::physical_plan::{collect, common};
-    use crate::prelude::SessionContext;
     use crate::test::exec::{
         assert_strong_count_converges_to_zero, BlockingExec, PanicExec,
     };
@@ -184,8 +183,7 @@ mod tests {
 
     #[tokio::test]
     async fn merge() -> Result<()> {
-        let session_ctx = SessionContext::new();
-        let task_ctx = session_ctx.task_ctx();
+        let task_ctx = Arc::new(TaskContext::default());
 
         let num_partitions = 4;
         let csv = test::scan_partitioned_csv(num_partitions)?;
@@ -212,8 +210,7 @@ mod tests {
 
     #[tokio::test]
     async fn test_drop_cancel() -> Result<()> {
-        let session_ctx = SessionContext::new();
-        let task_ctx = session_ctx.task_ctx();
+        let task_ctx = Arc::new(TaskContext::default());
         let schema =
             Arc::new(Schema::new(vec![Field::new("a", DataType::Float32, true)]));
 
@@ -235,8 +232,7 @@ mod tests {
     #[tokio::test]
     #[should_panic(expected = "PanickingStream did panic")]
     async fn test_panic() {
-        let session_ctx = SessionContext::new();
-        let task_ctx = session_ctx.task_ctx();
+        let task_ctx = Arc::new(TaskContext::default());
         let schema =
             Arc::new(Schema::new(vec![Field::new("a", DataType::Float32, true)]));
