+++ datafusion/core/src/physical_plan/stream.rs
@@ -339,11 +339,8 @@ mod test {
     use super::*;
     use arrow_schema::{DataType, Field, Schema};
 
-    use crate::{
-        execution::context::SessionContext,
-        test::exec::{
-            assert_strong_count_converges_to_zero, BlockingExec, MockExec, PanicExec,
-        },
+    use crate::test::exec::{
+        assert_strong_count_converges_to_zero, BlockingExec, MockExec, PanicExec,
     };
 
     fn schema() -> SchemaRef {
@@ -382,8 +379,7 @@ mod test {
 
     #[tokio::test]
     async fn record_batch_receiver_stream_drop_cancel() {
-        let session_ctx = SessionContext::new();
-        let task_ctx = session_ctx.task_ctx();
+        let task_ctx = Arc::new(TaskContext::default());
         let schema = schema();
 
         // Make an input that never proceeds
@@ -408,8 +404,7 @@ mod test {
     /// `RecordBatchReceiverStream` stops early and does not drive
     /// other streams to completion.
     async fn record_batch_receiver_stream_error_does_not_drive_completion() {
-        let session_ctx = SessionContext::new();
-        let task_ctx = session_ctx.task_ctx();
+        let task_ctx = Arc::new(TaskContext::default());
         let schema = schema();
 
         // make an input that will error twice
@@ -440,8 +435,7 @@ mod test {
     ///
     /// panic's if more than max_batches is seen,
     async fn consume(input: PanicExec, max_batches: usize) {
-        let session_ctx = SessionContext::new();
-        let task_ctx = session_ctx.task_ctx();
+        let task_ctx = Arc::new(TaskContext::default());
 
         let input = Arc::new(input);
         let num_partitions = input.output_partitioning().partition_count();
