+++ datafusion/core/tests/sqllogictests/test_files/options.slt
@@ -1,0 +1,84 @@
+# Licensed to the Apache Software Foundation (ASF) under one
+# or more contributor license agreements.  See the NOTICE file
+# distributed with this work for additional information
+# regarding copyright ownership.  The ASF licenses this file
+# to you under the Apache License, Version 2.0 (the
+# "License"); you may not use this file except in compliance
+# with the License.  You may obtain a copy of the License at
+
+#   http://www.apache.org/licenses/LICENSE-2.0
+
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an
+# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+# KIND, either express or implied.  See the License for the
+# specific language governing permissions and limitations
+# under the License.
+
+#######
+## Tests for config options
+#######
+
+
+statement ok
+create table a(c0 int) as values (1), (2);
+
+# Expect coalesce and default batch size
+query TT
+explain SELECT * FROM a WHERE c0 < 1;
+----
+logical_plan
+Filter: a.c0 < Int32(1)
+--TableScan: a projection=[c0]
+physical_plan
+CoalesceBatchesExec: target_batch_size=8192
+--FilterExec: c0@0 < 1
+----MemoryExec: partitions=4, partition_sizes=[1, 0, 0, 0]
+
+##
+# test_disable_coalesce
+##
+
+statement ok
+set datafusion.execution.coalesce_batches = false
+
+# expect no coalsece
+query TT
+explain SELECT * FROM a WHERE c0 < 1;
+----
+logical_plan
+Filter: a.c0 < Int32(1)
+--TableScan: a projection=[c0]
+physical_plan
+FilterExec: c0@0 < 1
+--MemoryExec: partitions=4, partition_sizes=[1, 0, 0, 0]
+
+statement ok
+set datafusion.execution.coalesce_batches = true
+
+
+##
+# test_custom_batch_size
+##
+
+statement ok
+set datafusion.execution.batch_size = 1234;
+
+# expect batch size to be 1234
+query TT
+explain SELECT * FROM a WHERE c0 < 1;
+----
+logical_plan
+Filter: a.c0 < Int32(1)
+--TableScan: a projection=[c0]
+physical_plan
+CoalesceBatchesExec: target_batch_size=1234
+--FilterExec: c0@0 < 1
+----MemoryExec: partitions=4, partition_sizes=[1, 0, 0, 0]
+
+
+statement ok
+set datafusion.execution.batch_size = 8192;
+
+statement ok
+drop table a
